Fix blessed mapping covering seen mons with objs
This fixes a bug where reading/casting blessed magic mapping would
obscure monsters otherwise in sight with detected objects that they
carried for a turn.  This doesn't happen in UnNetHack for some reason.

This is meant to go with commit 6c9adc7798 (CS989 improve non-cursed
scrolls of magic mapping).

Fixes #72 on GitHub.