NH4: Make level mon gen independent of player lvl
Adapted to work with the UnNetHack changes to monster difficulty.

It seems this change was made to support the (future) elimination of
object migration by sending objects straight to their target level,
which requires that level to be created on-the-fly.

make level generation independent of the player's level + depth
Daniel Thaler <daniel@dthaler.de>
98a3b471ab8432680b5be886adc208aaf25bd867