NH4: Eliminate object migration
migrating_objs is no more.  Also, calls to add_to_migration() are
replaced with the new deliver_object() function; see the code for
examples of how to do this conversion.

Of note, deliver_object() creates a level for objects to arrive on
on-the-fly if they don't already exist, explaining the splitting of
level generation from character experience level earlier.  Why object
migration was changed at all is a mystery to me, but not porting it
would mean make the porting of NetHack4 save code in the future even
harder than it will already be.

eliminate object migration
Daniel Thaler <daniel@dthaler.de>
ed8960f7eb32d2ea427ae8663b63d83951145900