Fix "You hear The knife tumbles downwards."
Without this fix, this kind of message appears when dropping an item
over a pit while blind, complete with bad capitalization of "The" and
"tumbles" instead of "tumble".  Found by boxo/Box_ in the #unnethack IRC
channel on FreeNode.

This bug also exists in vanilla NetHack, upstream NitroHack, UnNetHack
and NetHack4.