Added weeping angels.
Weeping angels possess three melee attacks, all affected by magic
cancellation: a level draining attack, a magic draining attack, and a
levelporting attack. In the source mythology, they are "creatures of
the abstract" that feed off both energy and the potential lives of
their victims, the latter obtained by transporting them back in time.
The three combined attacks are the (Un)NetHack equivalent of this. (The
levelporting attack is not yet enabled for weeping angels hitting
other monsters. The current excuse is that only the player has much of
a potential life to feed upon.)

They also possess a gaze attack. In the source mythology, looking into
the eyes of a weeping angel creates an angel within your mind. If the
player is blind, they cannot see the angel's eyes, and if they have
reflection, the angel avoids looking at them in order that it not
observe itself. Fighting the mental reflection of an angel, although
dangerous, exercises intelligence. One should be careful, however, as
it also does flat damage without regard for AC or resistances.

Also added is a weeping angel themed big room.

git-svn-id: https://svn.code.sf.net/p/unnethack/code/trunk@1293 f1e491ac-5ded-4967-80bf-9f4855be876a