Fix replay desync warning due to timer_id
The ID for timers is assigned from an incrementing static int which
wasn't reset, saved or restored with games, causing the command-driven
replays to use different IDs from the ones in the logged save diffs, in
turn causing a desync warning to be displayed.

This bug also exists in NetHack4.

Closes #77 on GitHub.