NH4: Don't call check_special_room on game restore
From NetHack4:

"As far as I can tell, it's a no-op; there isn't a matching call in
vanilla, and all it does is set urooms0 to urooms.  urooms0 is ignored
everywhere except inside check_special_room anyway; and saving and
restoring shouldn't cause you to leave or enter a room, so nothing
should happen.  The change to urooms0 confuses the desync detector, so
it's better to remove this call than leave it in."

Don't call check_special_room on game restore
Alex Smith <ais523@bham.ac.uk>
60454da37c4eccb920ee4a241474e31febace68b