Save and restore u.ubirthday to fix replays
Without this, restoring a game or rewinding in a replay would cause
u.ubirthday to be set to 0 instead of time the character was created.
This affects:

 * which bones file is chosen to be saved or loaded in the current
   level's bones pool
 * which ant type occupies any given anthole
 * the randomized creator deity name
 * the erosion applied to T-shirt labels when reading them
 * the cost of gems in a shop
 * which name is randomly chosen for shopkeepers

The last one was triggering a diff log replay desync warning when
rewinding a replay and advancing to a level a shopkeeper was created on,
since the name chosen by the replay was different from the one in the
logged diff.  The anthole population and gem costs would almost
certainly cause game-affecting desyncs as well.

This fix assumes 64 bits is enough to store time_t, since the C99
standard (ISO/IEC 9899/1999, n1256.pdf) states "the range and precision
of times representable in clock_t and time_t are implementation-
defined."  (It's 4 bytes large on my 32-bit Linux system.)  NitroHack
already assumes this via storing the result of time(NULL) in the
unsigned long long turntime variable.

The character's birthday is already stored in the log header, but due to
the heavily-stateful nature of log reading and the mid-log recovery of
the new logged diff replay system, it is easier and more reliable to
save and restore it with the rest of struct you.

NetHack4 partially addresses this in the following commits (code not
used here):

 * bed7cae11304dd0d5c2679c9f2bc5461cc3fb877
 * f893eb81fb405bcbe8ae623729351245f06cfd77

However, they only fix this for manually-loaded games, not for rewinding
in replays, so this bug also exists in NetHack4.