NH4: Don't load whole save files during replays
Before this, NitroHack would load the entire save log into memory to
split it into tokens when showing a replay.  The new save diff logging
from NetHack4 adds the new diff log entry type, which are large compared
to other types even with compression; I think this is why NetHack4
replaces this tokenization scheme with more direct file access seen in
this commit.

From NetHack4:

"This is half of what's necessary to make loading saved games much
faster.  (The other half is not scanning the entire file for options.)
Hopefully eventually, only the bits of the save file that are actually
relevant will ever be read."

With this change, the max action count in a replay must be stored in the
log header, replacing the string "NitroHack" ("NetHack4").  This needs
to be updated in log_command_result() which lives in log.c, but can only
be initialized properly when the log header is read: replay_begin() in
logreplay.c in this case.  NetHack4 and hence this code transports this
with the action_count global variable (yuck).  For future improvement,
I'd suggest altering replay_read_newgame() to grab this count since it
already reads the log header, and allowing the recently-added log_init()
function to accept the max action count so far, allowing action_count to
at least be static to log.c.

Don't store save files in memory during replays
Alex Smith <ais523@bham.ac.uk>
6b080354292807f249587fa4b7ab3b0bec650620