Cygwin: Note libmpfr4 dependency for gcc4
At some point in the past, the libmpfr4 dependency for gcc4 was botched
by its maintainer, but they found and fixed it.  However, some Cygwin
package repositories still miss it, so this commit adds a note to the
Cygwin build instructions to ensure it's there before proceeding.

Fixes issue #73 on GitHub.