Make sparkle option a number for animation speed
The sparkle option is now a number from 0 to 21 inclusive: 0 disables
the animation, 1 is the normal NetHack speed, 2 is double, etc. and 21
shows a single frame for a fraction of a second.  Based on the
sparkle_speed option from K-Mod.

flags.sparkle was changed from a boolean to a char, so it already has an
mwrite8() in save_flags() and an mread8() in restore_flags(); keep this
in mind if basing a patch off of this.