NH4: Split build date from build utilities
This allows makedefs to create verinfo.h for lev_comp and dgn_comp,
which is a version of date.h without the build date info.

Compared to NetHack4, this does not make the changes to support aimake
since this branch won't be using it.  The comment in nitrohack_types.h
claiming "CMake just sets UNIX, and only sometimes :)" is wrong; CMake
doesn't add any flags of its own when compiling.  The commit message
claims that this "fixes the problems with cmake where it would rebuild
all the files" but I'm not sure what this refers to specifically.

Build system updates (cmake and aimake)
Alex Smith <ais523@bham.ac.uk>
43a4cbf71040d9156faa38aa062e0c4b04429918