Don't "darken" extended terminal colors
It turns out that the "(color - 8) == darker" thing only applies to the
first 16 colors of the standard terminal color palette; other colors
turn into all sorts of strange shades when altered like this, so disable
this for those additional colors.

This is meant to go with commit ae95528b99 (NH4: configurable terminal
color mapping).