NH4: Comma selects ground for eat/tin/sacrifice
Compared to NetHack4, the prompts for ground objects won't show the
object picking prompt again if 'n' is given for all of them, the object
class groups avoid having to deal with unexplained "+ 2" insertions in
unrelated code, and an explanation for the goto is given in a
descriptive comment.  The code is also adapted slightly to work with the
Astral Escape patch included in UnNetHack.

Slash'EM-style eating controls
Alex Smith <ais523@bham.ac.uk>
28b83ea1b49ba2b06bbb6023d06db2dfc8ade5f6

Only skip the y/n dialog for corpses on the floor when eating.
nutzy <clcasto@gmail.com>
fda33426492a2fde36a8194cd276298e1a941a47

A better fix for sacrificing corpses on the floor
nutzy <clcasto@gmail.com>
322aa799b03577209b162bc72d9c175f0128c3b7