Pass right level to mksobj()
Without this, crashes can occur if mksobj() ever tries to use the lev
argument it's passed.  This fixes the mksobj calls in:

 * mongets()
 * mkmonmoney()
 * m_initinv()
 * mkcorpstat()
 * mk_mplayer_armor()
 * mkroll_launch()

Most of the fixes assume that objects made for monsters should be made
on the same level as that monster, which seems like a safe bet.

This bug also exists in upstream NitroHack and NetHack4.