oprops: Merge properties regardless of known-ness
... and merge the known properties correctly.

Without this fix, when several arrows of (unknown) fire miss and then
hit a monster, the ones that hit would have the fire property
identified, but not the ones that missed, and would thus not stack, even
if they were obviously originally from the same stack.

This does lead to players magically knowing the properties of a newly
found object if it stacks with an existing object with matching
properties that are known, but that's a quirk I can live with.