Don't allow running/travel when stunned/confused
Autoexplore already does this, but the code for it is now bundled with
this fix.

While we're at it, ensure flags.run is set to 0 when movement is blocked
in command_input() and when occupations are stopped or finished by
unmul().  Otherwise, normal movement could be stopped by stunning or
confusion when it shouldn't be.

This is related to commit a7f6b941a5 (interrupt running/travel/
autoexplore when confused).

This bug also exists in upstream NitroHack.

Fixes #83 on GitHub.