CS1353: autounlock when #looting locked containers
If a container is found to be locked when trying to open it and a
suitable unlocking tool is found, ask to unlock it.  Originally from
UnNetHackPlus.

The #loot command still needs to be used again to actually open the
container; ideally this would happen after the container is successfully
unlocked.

Compared to UnNetHack, answering 'y' to picking the container's lock
doesn't ask to loot to other containers on the same space.