Fix replay hiccup from port of autounlocking doors
The variable dz in pick_lock() is left uninitialized if non-zero rx and
ry arguments are given to the function which, when checked later in the
function, leads to desync between a game and its replay.

This is meant to go with commit ff361dbf02 (CS884 ask to unlock when
opening locked doors).