Autounlock #loots containers after unlocking
When the player chooses to autounlock a container when looting, the
looting menu will be shown once the container has been successfully
unlocked.  At long last, this makes locked containers as easy to loot as
unlocked ones; it's even better with NetHack4's 'a,' #loot shortcut.

Credit to Bulwersator for the idea.