NH4: Fix to show objects & traps on Plane of Water
This fixes the bug where objects and traps on the Plane of Water would
appear as water spaces from afar, due to the map_foo() functions calling
dbuf_set() based on the hero memory of spaces, which isn't updated on
the Plane of Water.

This does not fix a related bug where the same objects and traps are
still described as water in description messages.

Fix to show objects and traps on water level.
Jared Minch <jmminch@gmail.com>
c3381e381fd91913e5958302d2840d717650732e