Revert "Fix position descriptions on the Plane of Water"
This reverts commit ef24a8ae0b57dbbc68a28c3b9ebec93dcfe770c8.

This commit was causing objects with a randomized appearance to be
double obfuscated, so they'd have different descriptions when viewed
from afar and in the inventory.  This also broke the special description
logic for Sokoban prizes, i.e. true object name when seen and "a sokoban
foo" when not seen.

The problem that this reverted commit was trying to fix is that
nh_describe_pos() needs level memory to do its job, but the Plane of
Water doesn't have it.  I tried using the dbuf (display buffer) for this
no-level-memory case instead, but the dbuf is post-object-obfuscation,
and describe_object() used by nh_describe_pos() needs
pre-object-obfuscation data, otherwise the visible and actual object
types mismatch, causing describe_object() to create a fake object based
on the visible object type and use that to get the object name, which in
this case produces a completely incorrect name.

The real fix for this problem is not trivial, since nh_describe_pos()
and friends need (struct rm *) data, but for the Plane of Water that
data is created temporarily and disposed of once it's been used to
populate the dbuf, which is too late in the display pipeline to use.
The map_foo() functions compile this info and would be usable if they
didn't also reveal info, which is bad for a player-controllable cursor.
They could still be used if cansee() is identical to the vision rules on
the Plane of Water.