oprops: Fix artifact beheading, Mjollnir message
The beheading (and slicing) effect of Vorpal Blade and Thiefbane (and
the Tsurugi of Muramasa) was broken by the artifact_hit() refactoring.
Mjollnir was doing its damage, but the special hit message wasn't being
shown.

This is meant to go with commit d7d07e4d60 (oprops: magical object
properties from GruntHack).

This fixes #84 on GitHub.