CS319:Let One-eyed Sam wield Thiefbane with shield
I made a mistake last time and didn't realize that bhaak had
special-cased the two-handed Thiefbane for the Black Marketeer to wield
even with a shield, so her load-out is now the same as in UnNetHack
again:

 * Thiefbane
 * shield of reflection
 * magic dragon scale mail
 * speed boots
 * amulet of life saving
 * skeleton key

Compared to UnNetHack, this also handles the inverse case to permit
One-eyed Sam to wear a shield even when wielding a two-handed weapon.

This reverts commit 5522c346ee (change inv to let Black Marketeer wield
Thiefbane).

This is meant to go with commit 9dd9382aa9 (CS11-13,314,695: Black
Market by M. Campostrini).

This fixes #80 on GitHub.