oprops:Thiefbane decapitates 10% of the time again
The early check for a two-handed beheading artifact was eating the case
where the d20 hit roll was 1, so Thiefbane only decapitated on a roll
less than 3 and not 1, in other words only on 2.  This fix makes it
trigger on 1 and 2 again, like it should.

This is meant to go with commit d7d07e4d60 (oprops: magical object
properties from GruntHack).