Allow '^' for traps in targeting mode
They're technically not background elements, but it's strange to have
shortcuts for stairs and ladders but not magic portals.

I hard-coded the key because the alternative would be ten times more
complicated, and the targeting logic is already getting out of control
without adding more to it.

This closes #86 on GitHub.