NH4: Avoid a crash in monster ranged AI
This is more of a safety patch than a fix, since the crash doesn't occur
here; I'm guessing it's to do with AceHack/NetHack4's inclusion of the
Intelligent Pets patch.

Compared to NetHack4, this doesn't change spitmm() because that function
doesn't exist.

Avoid a crash in monster ranged AI
Alex Smith <ais523@bham.ac.uk>
3df54b4b405cf0aeaf42a596962a2f544b24274f