Fix C343-415: Mail daemons can be created when populating special levels with demons
Special level generation in NetHack and UnNetHack mask out G_NOGEN when
randomly creating monsters from a class, leading to mail daemons on
special levels that ask for '&' and chromatic dragons when they ask for
'D'.

This is inspired by a fix in GruntHack (also ported to NetHack4), but
replaces the hard-coding of Kops and eels with logic that checks that
all monsters of a class are flagged G_NOGEN before making an exception.

This replaces commit 0b268abae3 (respect G_NOGEN when making mon by
class on splevs).