Don't inverse detected monsters, remove use_inverse option
Using inverse/reverse video for detected monsters was ugly, and the only
reason for the use_inverse option.  I'd used that option previously for
some bar drawing stuff because I thought I needed to, but just now I
realized that A_REVERSE is used freely in other places, so there's no
need for it anymore.