More accurate cumulative heavyshot and artifact double damage
2x heavyshot damage coupled with 2x artifact damage should give 3x
instead of 4x, but since these damage modifiers are applied at different
places in the code, we add 1 to the multiplier and halve it to prepare
for the artifact code to do its damage doubling.

The issue that this commit solves is that the halving rounds when the
result is odd.  Previously, the code randomly chose for the end result
to be 1 higher or lower.  This commit replaces it by adding the
rounded-off part instead.

This is meant to go with commit f371c70783 (heavyshot: damage multiplier
instead of multishot).