Only quest nemesis can steal your quest artifact
In particular, your role's quest artifact can no longer be stolen by the
Wizard of Yendor, which is one of the biggest reasons NetHack players
almost always leave it behind.

The quest artifact is a special object with unique abilities that is
uniquely obtained by characters of each role, so it seems like an
obvious way to differentiate roles.