Reduce to-hit bonus from +Luck to rnd(Luck)
Having an invisible attribute contribute more to to-hit than dexterity
was silly: dex 18 gives +4, max dex (with +5 gauntlets of dexterity)
gives +9 (most characters) or +11 (dwarves), while max luck gave +13
(which will now be +d13 instead).

This should make the game a bit less of a cakewalk for veterans, and put
newer players at less of a disadvantage for not knowing how luck works.