Learn weapon enchantment on hit based on skill/race
Unskilled/basic/skilled/expert has a 1 in 1000/200/100/50 chance to
discover the enchantment of weapons, launchers and ammunition.  Weapons
of the hero's race double this chance.

Technically the chance is slightly lower than this, since the code that
does this relies on the valid_weapon_hit flag in hmon_hitmon(), which is
only set if the weapon hit does more than 1 damage.

This also fixes a couple of places where the weapon object considered by
hmon_hitmon() may become a dangling pointer, for safety's sake.