Choose containers from a menu when #looting
The ynq prompt is no more; instead, when looting over multiple
containers, players choose which containers to loot from a menu.

The paranoid_loot option now controls if a single container on a space
brings up a menu or skips straight to looting it.

This also causes the bite of a bag of tricks to stop looting of multiple
containers on a single space, as well an exploding bag of holding.
Autoloot after autounlock also stops multi-looting.

query_objlist() can now return -2 if the SIGNAL_ESCAPE flag is sent and
the player cancels the object menu (versus just not choosing anything),
allowing them to skip the adjacent monster loot prompt by pressing the
escape key.