Remove luck from to-hit, halve monk body armor penalty
Luck already affects a plethora of random events, so doing this puts
unspoiled players at less of a disadvantage for not understanding the
unintuitive mechanics surrounding luck.

Players who like body armor on their monks used luck to counter the -20
to-hit penalty, so reduce it to -10 to compensate for this change.

This is a follow-up to commit 30067d20cb (reduce to-hit bonus of luck).