Wrap rewinded messages like printed ones, clear empty lines
The message window becomes really ugly when rewinding during replays.
This commit was part of my attempt to solve this, but the truth is that
deep systemic changes are needed to allow rewinding to fill the message
history and window properly.

Not only does rewinding not prune messages or redraw the message window
like it should, it doesn't even have the data needed to prune the right
messages: replay steps go by fine-grained "action count" (tracked
internally and implicitly by the replay system), but messages are only
annotated with coarse-grained turn count.  Not to mention the message UI
code is a tangle of spaghetti logic and what are essentially global
variables.

So for now, rewinding the message window is now slightly less broken,
but still really ugly.