Dipping pure and diluted potions of the same type combines them
A stack of pure potions dipping into a stack of diluted potions of the
same type will now create a single stack of diluted potions.  This also
accounts for the blessings and curses of both stacks.

This reduces the backtracking created by water trips to and from potion
stashes leading up to and during alchemy sessions.