Bump patchlevel to 0.5.1
The `dynahack-0.5.0-win.zip` archive bundled compiled level files whose
version mismatched with the game actual, causing it to freak out and not
touch them, instead creating generic maze levels.

This packaging error is due to some sort of shortcoming in the way
Daniel Thaler asked CMake to compile level files (which was not amended
by changes made in NetHack4), causing them to fail to either be
recompiled individually or repackaged into `nhdat`.  This is a serious
bug that has laid dormant in the build system for a long time.

In the meantime, release builds should always be made with fresh `build`
and `install` directories.