Bump patchlevel to 0.5.2
This release greatly increases save file integrity and ability to
recover by fixing cases of incomplete data being logged that were
overlooked by NitroHack and NetHack4.

This version is save-compatible with v0.5.0 and v0.5.1, and players of
these versions are urged to update.  Save files for games that fail to
recover automatically can be sent to tungtn3+dynahack@gmail.com for
manual recovery.