Don't shop-donate non-empty bags dropped in sellobj_state SELL_DONTSELL
In NetHack source-diver jargon, containers belonging to the hero could
land in costly_spots under non-selling conditions without having their
no_charge flag set if said container was saleable() to the shop,
effectively donating the container (but not its contents) to the shop.

This could cause bags belonging to the hero to mysteriously become the
shopkeeper's property if they were to be dropped under
sellobj_state(SELL_DONTSELL).  Currently this state is only set for
partly-eaten food when eating is interrupted, so this doesn't happen in
practice.

In theory it could also happen if a zero-cost container owned by the
hero holding gold and other objects were dropped in a general store or
tool shop, but in practice there are no zero-cost containers, so nobody
ever noticed this.

This may be related to SC343-21: "Accounting is incorrect for containers
dropped in a shop that does not sell them."

This bug also exists in vanilla NetHack and likely all of its variants.