Credit/debit gold in containers even in sellobj_state SELL_DONTSELL
There's no capacity for the shop logic to handle gold without also
changing the credit/debit within it, so gold must always be handled in
sellobj, even when the state of it is set to SELL_DONTSELL.

This is required for an upcoming commit that will fix billing and credit
when nesting containers on shop floors.

This bug also exists in vanilla NetHack and likely all of its variants.