Fix billing/credit when hero nests their containers on a shop floor
This fixes a bug where the hero could accidentally donate the contents
of their bag to a shopkeeper if they put it into another bag on the shop
floor that also belonged to the hero.  To reproduce:

 1. Drop a sack on a shop floor, but don't sell it.
 2. Get another sack and put in non-shop-owned items.
 3. Put the sack with items into the sack on the shop floor.
 4. Take out the sack with items from the sack on the shop floor.

The shopkeeper will claim you owe them for the items in the sack, and
viewing the contents of the sack will show them belonging to the
shopkeeper.

This bug also exists in vanilla NetHack and likely all of its variants.