Bill/credit tipped container contents as if they had been picked up
This fixes shopkeepers buying their own goods when containers were
tipped on their shop floors (amongst other issues), by applying the
special logic for objects looted from containers on shop floors to the
contents of tipped containers.

Credit to guest41 and UnNetHack r1852 for spotting the issue, prompting
me to research the logic of containers in shops, leading to this fix for
this branch's #tip command, plus several fixes for bugs with containers
in shops in vanilla NetHack featured in previous commits.

Major findings of the shop logic research:

 * Objects on shop floors belong to the shopkeeper, unless they're
   flagged no_charge, which is used to mark objects the player drops
   but chooses not to sell or the shopkeeper doesn't accept.

 * Picking up a shop object adds it to the shopkeeper's bill and sets
   the object's unpaid flag, unless the object was flagged no_charge.

 * Regardless of billing/unpaid, all picked up objects have their
   no_charge flag unset.  Drop-to-sell only checks billing/unpaid, and
   not the no_charge flag.

 * All of the above bookkeeping applies recursively to container
   contents as well, even when not buying or selling, e.g. dropping but
   not selling a bag in a shop will flag the bag and all of its contents
   as no_charge, while picking up a bag in a shop will bill for all
   contents not flagged no_charge.

 * Gold always adds to credit/debit when dropped and takes away from
   credit/debit when picked up, even if the gold is in a container, or
   the player isn't buying/selling anything.

This is meant to go with commit 2e1b9c5787 (add new #tip command).