The frequency of Rodney's harassment of the player now slightly depends on luck.
If the player is at +8 luck, the rate is the same as vanilla. If your luck is above +8, the rate is slightly slower than vanilla, and if it is below +8, it is faster than vanilla. 

This is intented as a temporary change to the behavior, to be eventually replaced by the extended Rodney patch, once that is coded.

git-svn-id: https://svn.code.sf.net/p/unnethack/code/trunk@1876 f1e491ac-5ded-4967-80bf-9f4855be876a