Demaze Gehennom: Remove Heck^2 levels
Pretty much all of these levels are between 60-95% maze filler, which
are almost all less interesting than the current hellfill lava caverns.

The exceptions are the 4 new demon lord lairs which other variants have
better versions of, heck-b-4 (Lava2) which is just a lava cavern level
minus walls, and heck-d-8 (Pillars) which is essentially a Gehennom big
room and may be worth revisiting in the future.