Demaze Gehennom: Remove mazes around Wizard's Tower
The exterior of the Wizard's Tower levels and Fake Wizard's Tower
entrances are now lava caverns.

The fixed level pieces needed to be adjusted to not cause parts of the
level to be unreachable from each other.

Also, NOMAP inside a FUNCTION causes lev_comp to segfault when compiling
the game, so it's placed outside instead.