Demon lairs: Demogorgon, Geryon, Dispater, Yeenoghu (GruntHack)
This adds to Gehennom demon lairs for (in descending order of likely
depth): Yeenoghu, Dispater, Geryon and Demogorgon.

Demogorgon definitely earns his title of hardest enemy in NetHack.
DynaHack doesn't let the player get as overpowered as in GruntHack, so
now that Demogorgon is guaranteed, there needs to be better ways to deal
with him.