Show messages for objects that are randomly cursed
Objects will "glow brown" or "glow with a black aura" if they're
randomly cursed by e.g. the Wizard of Yendor or by sitting on a throne
with bad luck, just like the message given when dipping it into unholy
water.

As a bonus the object's BUC status will be revealed as well, again like
unholy water.