Wearing armor/jewelry/eyewear reveals curses
Based on the same change in NetHack4.

The weirdness about how discover_cursed_equip() is called is to avoid
showing "deathly cold" messages on the very first turn at the beginning
of a new game, e.g. for the convict's cursed striped shirt.  The
Foo_on() functions for multi-turn equipment can check if they're past
the first turn to show the message.  Sub-turn armor runs in the same
turn as the 'W'ear action and can't make that distinction, so the 'W'ear
command code handles it instead.  Amulets and rings don't get equipped
by the new game code and can just go ahead and show the message.