Fix C343-146: Wand of striking discovered if unseen mon zaps door
Exact bug description: "Unseen wand of striking becomes known if zapped
by unseen monster and it hits a door."

A related bug not fixed by this commit is that the same thing happens if
a wand of striking hits objects.  The fix for that would be much more
complicated, since the data on if the zapping monster can be seen
(zap_oseen) is local to muse.c, but the callback used when objects get
affected by the zap exists in zap.c.