Deinstapetrify: delayed petrify when feeling cockatrice corpses
Feeling a location with a cockatrice corpse while blind or sacrificing a
cockatrice without gloves will start delayed petrification instead of
instapetrifying.

The first attempt of the relevant action will be interrupted if contact
causes delayed petrification or polymorphs the player into a stone
golem, but further attempts will work as normal.  This allows a blind
adventurer that feels a cockatrice corpse in an item pile to pick up a
lizard corpse in that same pile, or retry sacrificing a cocktrice
corpse to reduce prayer timeout.