Deinstapetrify: delayed petrify if polymorphed while riding cockatrice
Riding on a cockatrice with petrify resistance and losing it by being
polymorphed starts delayed petrification instead of instapetrifying.

Note: It's impossible to saddle a cockatrice or chickatrice (the only
monsters in the game that petrify on touch) anyway since they're too
small, let alone ride them, so this case will never be encountered in
gameplay.