Prevent random chance of same-race polymorph with wizmode #youpoly
This removes the 1-in-5 chance to have level/attributes shuffled instead
of getting the polymorph form you asked for in debug mode.

Seriously, this omission getting me killed in wizmode was driving me up
the wall when testing changes that required me to be in the form of
other monsters, like the last couple of commits.