Deinstapetrify: delayed petrify two-weapon with off-hand corpse
Beginning two-weapon combat with a cockatrice corpse in your off-hand
starts delayed petrification instead of instapetrifying.

In practice the game never permits two-weapon combat with a non-weapon
off-hand object ("Your cockatrice corpse isn't a weapon") and never gets
far enough to trigger this case, so it's impossible for this to happen
in gameplay.