Deinstapetrify: delayed petrify when mount changes into petrifying form
Riding on a monster and having it polymorph into a cockatrice starts
delayed petrification instead of instapetrifying.

In practice, when it turns into a cockatrice or chickatrice, it will
always lose its saddle and the player will always be 'safely' thrown off
i.e. they never come in petrifying contact with it, and therefore this
case will never occur in gameplay.