Update changelog for DynaHack 0.5.0
Quite a big release: 800 of the 1100 lines of the changelog are
dedicated to DynaHack 0.5.0!

Having a changelog like this has a bunch of nice effects:

 1. People can work out how DynaHack differs from other variants.

 2. People can follow changes without having to figure out the stream of
    consciousness that is the commit log.

 3. People who recommend DynaHack to others have something to show them.

 4. People coming from NetHack have something to consult in the absence
    of wiki spoilers so they don't get screwed over by changes they
    couldn't have possibly have known about otherwise.

I was afraid I'd been writing too much in these commit messages, but
without them putting together a changelog like this would have been
impossible, so I guess it worked out in the end.