Fix sustain ability/weak from hunger strength exploit
The -1 to strength when becoming weak from hunger could be blocked with
a ring of sustain abilities, allowing, e.g. a knight to go from 15
strength to 18/25 using their starting stacks of apples and carrots.

This commit fixes this exploit by allowing changes to strength due to
weakness from hunger to ignore sustain abilities, which happens when
eating or prayer fixes starvation.

For completeness and future-proofing, the starting attribute tweaks are
also forced in the same way.

Bug discovered by cschellenger on the #unnethack FreeNode IRC channel.

Fixes #90 on GitHub.