Unhardcode Sokoban prizes
This moves the creation of Sokoban prizes out of fixup_special() code
and into the Sokoban level descriptions themselves.  This is done by
adding a new `prize` flag for `OBJECT` statements to the lev_comp
grammar which sends `SP_O_V_PRIZE` to the special level coder, in turn
setting the object's `oprize` flag (really `olocked` overloaded).

Object naming has been adjusted to allow for potentially any object to
be obfuscated out-of-sight and identified in-sight.

Constants and struct fields dealing with prizes that contain some
variation of the name "Sokoban" have been renamed to be more general.