Move Sokoban prizes to Mines End
The prize at the end of Sokoban is now a sack with four random wands and
rings; the randomly-placed wands and rings on each Sokoban level have
been removed as a result.

This makes Sokoban less valuable and deeper Gnomish Mines more valuable,
making Sokoban now truly optional instead of optional-but-not-really.

This also guarantees a sack for characters unfortunate enough to not
find any bags in the entire dungeon.