Reduce/simplify demon bribe amount to 1000 to 2000
Demon bribery amount no longer depends on the amount of gold in your
open inventory, whether or not your alignment matches that of the demon,
or whether or not the demon is in its "home level".

The amount was reduced because though gold is not hard to find in
Gehennom, reserving inventory weight for it in the late game has always
been a pain, so players shouldn't feel forced to carry too much of it.