New ravine entrance level to Vlad's Tower (dNetHack)
Entering Vlad's Tower now presents an open ravine level with trees,
water, vampires, zombies, ravens and a drawbridge entrance leading into
the tower proper.  This makes the branch 4 levels long instead of 3.

This also adds a failsafe for positioning the player on new maps (the
lack of a down ladder in the previous bottom level of Vlad's Tower led
to very odd game-ending behavior when going up the ladder in the new
entrance level, which had to be fixed).