oprops: ID stealth on weapons on hit, put "silently" in hit messages
Hitting a monster with a weapon of stealth identifies the stealth
property on that weapon, and also gives a message that you "silently
hit" or "silently kill" the monster, hinting at how weapons of stealth
bypass the usual combat noise mechanic.

I wish I could put "silently" for other situations that noisy_hit()
allows to be silent, e.g. stealthy characters hitting with knives and
thrown darts/shuriken, but lining up the hit messages with the random
rolls in noisy_hit() would be a herculean task so it's not happening.